Overhaul CSS
The design was too wide, which meant that everything felt very fat
and bloated. I've slimmed the max-width down to a svelte 1024px, given
the main box a gentle blocky shadow, and adjusted margins.

The index page gets a layout change to remove the URL from the header
and introduce an accented horizontal rule. The "minimal buttons" are
gone and turned into a classic unordered list, replete with list-style-
type intact for a retro feel.